requirements.txt
# requirements.txt
# Dépendances principales pour l'API FastAPI + DB

fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1